refactor(site): extract soil conversion math into a tested module
The meq/l to PPM conversion, lab-range tags, and Four Lows detection
lived inside parse-soil-test.ts alongside its DB and CLI code, so the
pure math had no test coverage. Move the constants, types, and the
computeTag/buildMinerals/checkFourLows functions into
scripts/lib/soil-conversion.ts and cover them with unit tests. No
behaviour change to the script.
